Big UI overhaul
Removed big info menu on left. Symbolic bar on the bottom that conveys all that info. Country icons are changed to different polygons to represent shape. This is still color blind friendly and the markings no longer obstruct the text. Reinforcement now shows up as the same shape as the country's polygon.